+++ R/bru.inference.R
@@ -1311,6 +1311,8 @@ extended_bind_rows <- function(...) {
 #'   defaults to `FALSE`, unless `response_data` is non-`NULL`, or `data` is a
 #'   `list`, or the likelihood construction requires it.
 #' @param control.family A optional `list` of `INLA::control.family` options
+#' @param tag character; Name that can be used to identify the relevant parts
+#' of INLA predictor vector output, via [bru_index()].
 #' @param options A [bru_options] options object or a list of options passed
 #' on to [bru_options()]
 #' @param .envir The evaluation environment to use for special arguments (`E`,
@@ -1340,6 +1342,7 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
                  allow_latent = deprecated(),
                  allow_combine = NULL,
                  control.family = NULL,
+                 tag = NULL,
                  options = list(),
                  .envir = parent.frame()) {
   options <- bru_call_options(options)
@@ -1722,7 +1725,8 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
       domain = domain,
       used = used,
       allow_combine = allow_combine,
-      control.family = control.family
+      control.family = control.family,
+      tag = tag
     ),
     class = "bru_like"
   )
@@ -1730,6 +1734,118 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
   # Return likelihood
   lh
 }
+
+#' @title Extract predictor index information
+#'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#' Extract the index vector for a [like()] predictor,
+#' or the whole or a subset of a full [bru()] predictor.
+#'
+#' @param object A [bru()] or [like()] output object
+#' @param \dots Arguments passed on to sub-methods.
+#' @returns An `integer` vector.
+#' @examplesIf bru_safe_inla()
+#' \dontrun{
+#'    fit <- bru(
+#'      ~ 0 + x,
+#'      like(
+#'        y ~ .,
+#'        data = data.frame(x = 1:3, y = 1:3 + rnorm(3)),
+#'        tag = "A"
+#'      ),
+#'      like(
+#'        y ~ .,
+#'        data = data.frame(x = 1:4, y = c(NA, NA, 3:4) + rnorm(4)),
+#'        tag = "B"
+#'      )
+#'    )
+#'    bru_index(fit)
+#'    bru_index(fit, "A")
+#'    bru_index(fit, "B")
+#'    bru_index(fit, c("B", "A"))
+#'    bru_index(fit, what = "missing")
+#' }
+#' @export
+bru_index <- function(object, ...) {
+  UseMethod("bru_index")
+}
+
+#' @describeIn bru_index Extract the index vector for the predictor vector
+#' for a [like()] sub-model. The indices are relative to the sub-model, and need
+#' to be appropriately offset to be used in the full model predictor.
+#' @param what `character` or `NULL`; One of `NULL`, "all", "observed", and
+#'   "missing". If `NULL` (default) or "all", gives the index vector for the
+#'   full sub-model predictor. If "observed", gives the index vector for the
+#'   observed part (response is not `NA`). If "missing", gives the index vector
+#'   for the missing part (response is `NA`) of the model.
+#' @export
+bru_index.bru_like <- function(object, what = NULL, ...) {
+  what <- match.arg(what, c("all", "observed", "missing"))
+  size <- bru_response_size(object)
+  idx <- seq_len(size)
+  if (identical(what, "all")) {
+    return(idx)
+  }
+  resp <- object[["response_data"]][[object[["response"]]]]
+  if (is.vector(resp)) {
+    miss <- is.na(resp)
+  } else {
+    # Should work for both inla.surv and inla.mdata objects:
+    miss <- is.na(resp[[1]])
+  }
+  if (identical(what, "observed")) {
+    return(idx[!miss])
+  }
+  return(idx[miss])
+}
+
+#' @describeIn bru_index Extract the index vector for "APredictor" for one or
+#'   more specified observation [like()] sub-models. Accepts any combination of
+#'   `tag` and `what`.
+#' @param tag `character` or `integer`; Either a character vector identifying
+#'   the tags of one or more of the [like()] observation models, or an integer
+#'   vector identifying models by their [bru()] specification order. If `NULL`
+#'   (default) computes indices for all sub-models.
+#' @export
+bru_index.bru <- function(object, tag = NULL, what = NULL, ...) {
+  if (is.null(tag)) {
+    tag <- seq_len(length(object[["bru_info"]][["lhoods"]]))
+  }
+  if (length(tag) == 0L) {
+    return(integer(0))
+  }
+  size <- bru_response_size(object)
+  off <- c(0L, cumsum(size))
+  if (is.character(tag)) {
+    if (!all(tag %in% names(object[["bru_info"]][["lhoods"]]))) {
+      stop(paste0(
+        "Invalid tag(s) '", paste(tag, collapse = ", "), "' for ",
+        "bru object with tags '", paste(names(object[["bru_info"]][["lhoods"]]),
+                                        collapse = ", "), "'"
+      ))
+    }
+    unlist(lapply(tag, function(x) {
+      x_idx <- which(names(object$bru_info$lhoods) %in% x)
+      off[x_idx] + bru_index(object$bru_info$lhoods[[x_idx]], what = what)
+    }))
+  } else {
+    ok_tags <- (tag >= 1L) &
+      (tag <= length(object[["bru_info"]][["lhoods"]]))
+    if (any(!ok_tags)) {
+      stop(paste0(
+        "Invalid tag indices '", paste(tag[!ok_tags], collapse = ", "),
+        "' for bru object with tag indices '1, ..., ",
+        length(object[["bru_info"]][["lhoods"]]), "'"
+      ))
+    }
+    unlist(lapply(tag, function(x) {
+      off[x] + bru_index(object$bru_info$lhoods[[x]], what = what)
+    }))
+  }
+}
+
+
 
 
 #' @title Response size queries
@@ -1822,6 +1938,17 @@ like_list.list <- function(object, envir = NULL, ...) {
 
   class(object) <- c("bru_like_list", "list")
   environment(object) <- envir
+  names(object) <- vapply(
+    object,
+    function(x) {
+      if (is.null(x[["tag"]])) {
+        NA_character_
+      } else {
+        x[["tag"]]
+      }
+    },
+    ""
+  )
   object
 }
 
@@ -1919,7 +2046,8 @@ summary.bru_like <- function(object, verbose = TRUE, ...) {
       data_class = class(object[["data"]]),
       response_class = class(object[["response_data"]][[object[["response"]]]]),
       predictor = deparse(object[["formula"]]),
-      used = object[["used"]]
+      used = object[["used"]],
+      tag = object[["tag"]]
     ),
     class = "summary_bru_like"
   )
@@ -1947,12 +2075,14 @@ print.summary_bru_like <- function(x, ...) {
   cat(sprintf(
     paste0(
       "  Family: '%s'\n",
+      "    Tag: %s\n",
       "    Data class: %s\n",
       "    Response class: %s\n",
       "    Predictor: %s\n",
       "    Used components: %s\n"
     ),
     lh$family,
+    paste0("'", lh$tag, "'", collapse = ", "),
     paste0("'", lh$data_class, "'", collapse = ", "),
     paste0("'", lh$response_class, "'", collapse = ", "),
     if (length(lh$predictor) > 1) {
@@ -2063,6 +2193,8 @@ bru_like_control_family.bru_like_list <- function(x,
     }
   } else {
     control.family <- lapply(x, bru_like_control_family)
+    # inla() requires a unnamed list of lists
+    names(control.family) <- NULL
   }
   control.family
 }